apps: Extract script to load apps data.
This helps us run the script independent of the URL path.
