+++ templates/corporate/apps.html
@@ -1,5 +1,5 @@
 {% extends "zerver/portico.html" %}
-{% set entrypoint = "landing-page" %}
+{% set entrypoint = "apps-download" %}
 
 {% set PAGE_TITLE = "Download the Zulip app for your device" %}
