+++ web/webpack.assets.json
@@ -51,6 +51,11 @@
         "./src/portico/landing-page",
         "./styles/portico/landing_page.css"
     ],
+    "apps-download": [
+        "./src/bundles/portico",
+        "./src/portico/apps_download",
+        "./styles/portico/landing_page.css"
+    ],
     "landing-page-hello": [
         "./src/bundles/hello",
         "./src/portico/hello",
